+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Upcoming
 * Fix (un)packarchive error with UTF-8 characters (#219)
+* Fix broken stdin when using reptor as a library (#226)
 * Add MCP Server
 
 # 0.32